# Add inline tags to commonly used functions