--- a/rustler/src/types/list.rs
+++ b/rustler/src/types/list.rs
@@ -56,6 +56,7 @@ impl<'a> ListIterator<'a> {
 impl<'a> Iterator for ListIterator<'a> {
     type Item = Term<'a>;
 
+    #[inline]
     fn next(&mut self) -> Option<Term<'a>> {
         let env = self.term.get_env();
         let cell = unsafe { list::get_list_cell(env.as_c_arg(), self.term.as_c_arg()) };
@@ -78,6 +79,7 @@ impl<'a> Iterator for ListIterator<'a> {
 }
 
 impl<'a> Decoder<'a> for ListIterator<'a> {
+    #[inline]
     fn decode(term: Term<'a>) -> NifResult<Self> {
         match ListIterator::new(term) {
             Some(iter) => Ok(iter),
@@ -96,6 +98,7 @@ impl<T> Encoder for Vec<T>
 where
     T: Encoder,
 {
+    #[inline]
     fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
         self.as_slice().encode(env)
     }
@@ -105,6 +108,7 @@ impl<'a, T> Decoder<'a> for Vec<T>
 where
     T: Decoder<'a>,
 {
+    #[inline]
     fn decode(term: Term<'a>) -> NifResult<Self> {
         let iter: ListIterator = term.decode()?;
         let res: NifResult<Self> = iter.map(|x| x.decode::<T>()).collect();
@@ -116,6 +120,7 @@ impl<T> Encoder for [T]
 where
     T: Encoder,
 {
+    #[inline]
     fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
         let term_array: Vec<NIF_TERM> = self.iter().map(|x| x.encode(env).as_c_arg()).collect();
         unsafe { Term::new(env, list::make_list(env.as_c_arg(), &term_array)) }
@@ -126,6 +131,7 @@ impl<T> Encoder for &[T]
 where
     T: Encoder,
 {
+    #[inline]
     fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
         let term_array: Vec<NIF_TERM> = self.iter().map(|x| x.encode(env).as_c_arg()).collect();
         unsafe { Term::new(env, list::make_list(env.as_c_arg(), &term_array)) }
@@ -135,6 +141,7 @@ where
 /// ## List terms
 impl<'a> Term<'a> {
     /// Returns a new empty list.
+    #[inline]
     pub fn list_new_empty(env: Env<'a>) -> Term<'a> {
         let list: &[u8] = &[];
         list.encode(env)
@@ -144,6 +151,7 @@ impl<'a> Term<'a> {
     /// See documentation for ListIterator for more information.
     ///
     /// Returns None if the term is not a list.
+    #[inline]
     pub fn into_list_iterator(self) -> NifResult<ListIterator<'a>> {
         ListIterator::new(self).ok_or(Error::BadArg)
     }
@@ -156,6 +164,7 @@ impl<'a> Term<'a> {
     /// ```elixir
     /// length(self_term)
     /// ```
+    #[inline]
     pub fn list_length(self) -> NifResult<usize> {
         unsafe { list::get_list_length(self.get_env().as_c_arg(), self.as_c_arg()) }
             .ok_or(Error::BadArg)
@@ -171,6 +180,7 @@ impl<'a> Term<'a> {
     /// [head, tail] = self_term
     /// {head, tail}
     /// ```
+    #[inline]
     pub fn list_get_cell(self) -> NifResult<(Term<'a>, Term<'a>)> {
         let env = self.get_env();
         unsafe {
@@ -183,6 +193,7 @@ impl<'a> Term<'a> {
     /// Makes a copy of the self list term and reverses it.
     ///
     /// Returns Err(Error::BadArg) if the term is not a list.
+    #[inline]
     pub fn list_reverse(self) -> NifResult<Term<'a>> {
         let env = self.get_env();
         unsafe {