@hansihe 2022-09-11
This looks like a very good idea to me.