@evnu 2022-11-11
Interesting. I would not expect `inline` to make a real difference. Can we explain why we see a difference in the performance?